# cmd/api/docs/swagger.yaml
basePath: /api
definitions:
...
  handler.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  models.NginxSite:
    properties:
      accessLog:
        type: string
      aliases:
        items:
          type: string
        type: array
      configPath:
        type: string
      createdAt:
        type: string
      domain:
        type: string
      enabled:
        type: boolean
      errorLog:
        type: string
      forceHttps:
        type: boolean
      id:
        type: integer
      index:
        type: string
      name:
        type: string
      proxy:
        type: boolean
      proxyPass:
        type: string
      rewrite:
        type: string
      root:
        type: string
      ssl:
        type: boolean
      sslCert:
        type: string
      sslKey:
        type: string
      updatedAt:
        type: string
    type: object
...
  ssl.CreateSSLRequest:
    properties:
      acme_client_id:
        type: integer
      domain:
        type: string
    required:
    - acme_client_id
    - domain
    type: object
host: localhost:8080
info:
  contact:
    email: support@etapanel.com
    name: EtaPanel Support
    url: https://github.com/EtaPanel-dev/EtaPanel
  description: EtaPanel 服务器管理面板后端API文档，提供系统监控、文件管理、Nginx配置、SSL证书管理等功能
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: EtaPanel API
  version: "1.0"
paths:
...
  /auth/acme/ssl:
    get:
      consumes:
      - application/json
      description: 获取所有SSL证书信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  items:
                    type: object
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 获取SSL证书列表
      tags:
      - SSL证书管理
    post:
      consumes:
      - application/json
      description: 通过ACME协议申请SSL证书
      parameters:
      - description: SSL申请信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ssl.CreateSSLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 申请成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 申请SSL证书
      tags:
      - SSL证书管理
...
  /auth/nginx/sites:
    get:
      consumes:
      - application/json
      description: 获取所有Nginx网站配置列表
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.NginxSite'
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 获取网站列表
      tags:
      - Nginx管理
    post:
      consumes:
      - application/json
      description: 创建新的Nginx网站配置
      parameters:
      - description: 网站配置信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.NginxSite'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "409":
          description: 域名已存在
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 创建网站
      tags:
      - Nginx管理
  /auth/nginx/sites/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定ID的网站配置
      parameters:
      - description: 网站ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 网站不存在
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 删除网站
      tags:
      - Nginx管理
    put:
      consumes:
      - application/json
      description: 更新指定ID的网站配置
      parameters:
      - description: 网站ID
        in: path
        name: id
        required: true
        type: integer
      - description: 网站配置信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.NginxSite'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 网站不存在
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 更新网站配置
      tags:
      - Nginx管理
  /auth/nginx/sites/{id}/toggle:
    post:
      consumes:
      - application/json
      description: 切换指定ID网站的启用状态
      parameters:
      - description: 网站ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 网站不存在
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 启用/禁用网站
      tags:
      - Nginx管理
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 输入Bearer token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...